Host hamakor logo on site
Before this commit, we referenced Hamakor logo on linux.org.il.
That site runs on http (not https) and including an image from
it is a slight compromise of our own site's security -- and it
causes browsers to mark it as "containing insecure data".

While at it, removed some whitespace in that logo.