Replaced file-saving thumbnails with data-uris
This is done as a temporary means to enable serving thunbnailed
images, when our combined use of
- Heroku for hosting, with its ephemeral file system
- WhiteNoise for static files, and its ManifestFileStorage which
  seems to prevent including dynamic files (or not. can't tell now)
- npm whatever for the local dev, which messes a little with all the statics
seems to make it difficult to do otherwise.

This way, the thumbnailers can read the sponsor logos from the
file system (i.e. from MEDIA_ROOT) and nobody needs to write files
during work, so some of the crazy is prevented.

This comes at a stupid price of processing per page view, not using
cache for images, and is generally horrible. But it gives us time
to come up with something sane.